+++ migration.go
@@ -1,9 +1,6 @@
 package goose
 
-import (
-	"fmt"
-)
-
+// Migration represents a single migration.
 type Migration struct {
 	// Type is the type of migration (SQL or Go).
 	Type MigrationType
@@ -11,17 +8,18 @@ type Migration struct {
 	Source string
 	// Version is the parsed version from the migration file name.
 	Version int64
-	// Empty is true if the migration was a no-op, but was still recorded in the database (unless no
-	// versioning is enabled).
+	// Empty is true if the migration is empty. For SQL migrations, this means the file contains no
+	// SQL statements. For Go migrations, this means the migration functions for Up and Down is nil.
 	Empty bool
-	// UseTx is true if the migration was applied in a transaction.
+	// UseTx is true if the migration is safe to run in a transaction.
 	UseTx bool
 }
 
 func (m *Migration) String() string {
-	return fmt.Sprint(m.Source)
+	return m.Source
 }
 
+// MigrationType is the type of migration (SQL or Go).
 type MigrationType string
 
 const (
