+++ version.go
@@ -1,3 +1,1 @@
 package goose
-
-// TODO(mf): implement this
